+++ src/main/java/com/kustaurant/kustaurant/evaluation/infrastructure/evaluation/EvaluationJpaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 public interface EvaluationJpaRepository extends JpaRepository<EvaluationEntity, Integer>{
-    boolean existsByUser_UserIdAndRestaurant_RestaurantId(Integer userId, Integer restaurantId);
+    boolean existsByUser_UserIdAndRestaurant_RestaurantId(Long userId, Integer restaurantId);
     List<EvaluationEntity> findByRestaurant_RestaurantIdAndStatus(Integer restaurantId, String status);
 
     Integer countAllByStatus(String status);
@@ -22,9 +22,9 @@ public interface EvaluationJpaRepository extends JpaRepository<EvaluationEntity,
             ELSE e.createdAt
         END DESC
     """)
-    List<EvaluationEntity> findSortedEvaluationsByUserIdDesc(@Param("userId") Integer userId);
+    List<EvaluationEntity> findSortedEvaluationsByUserIdDesc(@Param("userId") Long userId);
 
     List<EvaluationEntity> findByStatus(String status);
 
-    List<EvaluationEntity> findByUser_UserId(Integer userId);
+    List<EvaluationEntity> findByUser_UserId(Long userId);
 }
