+++ src/main/java/com/kustaurant/kustaurant/feedback/infrastructure/FeedbackJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface FeedbackJpaRepository extends JpaRepository<FeedbackEntity, Long> {
-    List<FeedbackEntity> findAllByUser_UserId(Integer id);
+    List<FeedbackEntity> findAllByUser_UserId(Long id);
 }
