+++ src/main/java/com/kustaurant/kustaurant/global/auth/argumentResolver/AuthUserArgumentResolver.java
@@ -1,7 +1,7 @@
 package com.kustaurant.kustaurant.global.auth.argumentResolver;
 
 import com.kustaurant.kustaurant.global.auth.session.CustomOAuth2User;
-import com.kustaurant.kustaurant.user.domain.enums.UserRole;
+import com.kustaurant.kustaurant.user.user.domain.enums.UserRole;
 import com.kustaurant.kustaurant.global.exception.exception.auth.UnauthenticatedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
@@ -37,14 +37,13 @@ public class AuthUserArgumentResolver implements HandlerMethodArgumentResolver {
         }
 
         /* ───────────── principal → userId 추출 ───────────── */
-        Integer userId;
+        Long userId;
         if (auth.getPrincipal() instanceof CustomOAuth2User oAuthUser) {        // 웹 세션
             userId = oAuthUser.getUserId();
-        } else if (auth.getPrincipal() instanceof Integer uid) {                // 앱 JWT
+        } else if (auth.getPrincipal() instanceof Long uid) {                // 앱 JWT
             userId = uid;
         } else {
-            throw new UnauthenticatedException(
-                    "지원하지 않는 principal 타입: " + auth.getPrincipal().getClass().getName());
+            throw new UnauthenticatedException("지원하지 않는 principal 타입: " + auth.getPrincipal().getClass().getName());
         }
 
         /* ───────────── 권한(ROLE) 추출 및 매핑 ───────────── */
