+++ src/main/java/com/kustaurant/kustaurant/global/auth/argumentResolver/AuthUserInfo.java
@@ -1,9 +1,9 @@
 package com.kustaurant.kustaurant.global.auth.argumentResolver;
 
-import com.kustaurant.kustaurant.user.domain.enums.UserRole;
+import com.kustaurant.kustaurant.user.user.domain.enums.UserRole;
 
 public record AuthUserInfo(
-        Integer id,
+        Long id,
         UserRole role
 ) {
 }
