+++ src/main/java/com/kustaurant/kustaurant/post/domain/PostLike.java
@@ -8,11 +8,11 @@ import java.time.LocalDateTime;
 public class PostLike {
 
     private Integer id;
-    private Integer userId;
+    private Long userId;
     private Integer postId;
     private final LocalDateTime createdAt;
 
-    public PostLike(Integer userId, Integer postId, LocalDateTime createdAt) {
+    public PostLike(Long userId, Integer postId, LocalDateTime createdAt) {
         this.userId = userId;
         this.postId = postId;
         this.createdAt = createdAt;
