+++ src/main/java/com/kustaurant/kustaurant/post/infrastructure/PostLikeJpaRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PostLikeJpaRepository extends JpaRepository<PostLikeEntity,Integer> {
-    Optional<PostLikeEntity> findByUser_UserIdAndPost_PostId(Integer userId, Integer postId);
+    Optional<PostLikeEntity> findByUser_UserIdAndPost_PostId(Long userId, Integer postId);
 
-    Boolean existsByUser_UserIdAndPost_PostId(Integer userId, Integer postId);
+    Boolean existsByUser_UserIdAndPost_PostId(Long userId, Integer postId);
 
     int countByPost_PostId(Integer postId);
 }
